Improve BB caching & chunk logic

+++ src/main/java/de/jpx3/intave/IntavePlugin.java
@@ -494,7 +494,9 @@ public final class IntavePlugin extends JavaPlugin {
 
   @Native
   public void clearCacheFiles() {
-    configurationService.deleteCache();
+    if(configurationService != null) {
+      configurationService.deleteCache();
+    }
     clearIntegrityGarbage();
   }
